requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
dj-database-url==3.0.1
Django==5.2.4
django-livereload-server==0.5.1
Faker==37.5.3
frozenlist==1.7.0
idna==3.10
multidict==6.6.4
mysqlclient==2.2.7
pillow==11.3.0
propcache==0.3.2
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.3
tornado==6.5.1
twilio==9.7.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
yarl==1.20.1
python-dotenv==1.0.0
